addressing Franks second phase review comemnts
